# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Webinar ApplicationAPI",
        "version": "0.1"
    },
    "paths": {
...
        "/api/user/addToCart": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "add covent to a user cart",
                "description": "add covent to a user cart",
                "operationId": "addCoventToCart",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "covent_id",
                                    "count"
                                ],
                                "properties": {
                                    "covent_id": {
                                        "type": "text"
                                    },
                                    "count": {
                                        "type": "integer"
                                    },
                                    "coupon_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/user/deleteCart": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "delete from the users cart",
                "description": "delete from the users cart",
                "operationId": "deleteCart",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "cart_id"
                                ],
                                "properties": {
                                    "cart_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user/updateCart": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "update the users cart",
                "description": "update the users cart",
                "operationId": "updateCart",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "cart_id",
                                    "count"
                                ],
                                "properties": {
                                    "cart_id": {
                                        "type": "integer"
                                    },
                                    "count": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user/showInvoice": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "show user invoices",
                "description": "show user invoices",
                "operationId": "show user invoice",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "number": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/page": {
            "post": {
                "tags": [
                    "Page"
                ],
                "summary": "return a page",
                "description": "you should choose page_link or page_id or none of them",
                "operationId": "return a page",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "page_slug": {
                                        "type": "text"
                                    },
                                    "page_link": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": {
        "passport": {
            "type": "oauth2",
            "description": "Laravel webinar passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://localhost/oauth/authorize",
                    "tokenUrl": "http://localhost/oauth/token",
                    "refreshUrl": "http://localhost/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}